feat(content analytics) #33380 : Improve error messaging when typos in CubeJS query are present (#34109)
### Proposed Changes
* Improved error messaging from the back-end in order to point out more
details related to potential errors with CubeJS queries. This PR is
meant to handle as many error situations as possible.
* The main goal is to get to the root cause of the problem as fast as
possible, without having to dive into the logs when something goes
wrong.
* For now, the Angular layer is truncating the error messages as soon as
it finds a `:` character in any error message. We'll take care of that
in a separate ticket:
  * #34094

This PR fixes: #33380