Issue 33902 images on block editor not rendering (#34129)
### **Problem 1**
The first issue mentioned was caused because the `render.vtl` file
(which is the one used by the lists) wasn't handling the image type when
recursing into nested content.

**Solution**
The solution was to modify that file and add missing block type handler
for `dotImage`, also `dotVideo` and `dotContent` type was added.

### **Problem 2**
The image wasn't being included in a bundle when push publishing because
the Story Block processing only iterated through top-level blocks
without recursing into nested content arrays within container blocks.

**Solution**
Added a `processBlocksRecursively()` function to traverse the
content/items recursively and add the nested contents as dependencies.


This PR fixes: #33902